requirements.txt
streamlit
pandas
numpy
pymongo
bcrypt
matplotlib
plotly
scipy
openpyxl
pyyaml
pyjwt
python-dotenv